ci: use signing transforms from mozilla-taskgraph
